Assignment 01/pyproject.toml
[project]
name = "assignment-01"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents[litellm]>=0.2.9",
    "python-dotenv>=1.1.1",
]
